Separate logic of regular and stake tx at txHistory
