+++ app/components/TxHistory/index.js
@@ -1,11 +1,11 @@
-import TxHistoryRow from "./TxHistoryRow";
 
 const TxHistory = ({ transactions = [], limit, overview, tsDate }) => (
   <>
     {transactions.map( (tx, index) => {
       if(limit && index >= limit) return;
       return (
-        <TxHistoryRow {...{ key: tx.txHash, overview, tx, tsDate }} />
+        <div></div>
+        // <TxHistoryRow {...{ key: tx.txHash, overview, tx, tsDate }} />
       );
     })}
   </>
