+++ app/components/shared/index.js
@@ -17,4 +17,5 @@ export { default as Subtitle } from "./Subtitle";
 export { default as FormattedRelative } from "./FormattedRelative";
 export { default as CreatePassPhrase } from "./CreatePassPhrase";
 export { default as UnsignedTx } from "./UnsignedTx";
+export { default as RegularTxRow } from "./RegularTxRow";
 export * from "./RoutedTabsHeader";
