+++ app/components/views/HomePage/tables/RecentTransactions.js
@@ -1,16 +1,15 @@
 // @flow
 import { home } from "connectors";
 import { DecredLoading } from "indicators";
-import TxHistory from "TxHistory";
 import { FormattedMessage as T } from "react-intl";
 import "style/Fonts.less";
 import "style/HomePage.less";
 import NoTransactionsLinks from "./NoTransactionsLinks";
+import { RegularTxRow } from "shared";
 
 const RecentTransactions = ({
   transactions,
   getTransactionsRequestAttempt,
-  getAccountsResponse,
   rowNumber,
   goToTransactionHistory,
   tsDate
@@ -30,8 +29,13 @@ const RecentTransactions = ({
           }
         </div>
         <div className="home-content-nest">
-          {transactions.length > 0 ?
-            <TxHistory overview limit={rowNumber} {...{ getAccountsResponse, transactions, tsDate }} /> :
+          { transactions.length > 0 ?
+            transactions.map( (tx, index) => {
+              if(index >= rowNumber) return;
+              return (
+                <RegularTxRow key={tx.txHash} {...{ overview: true, tx, tsDate }} />
+              );
+            }) :
             <NoTransactionsLinks />}
         </div>
       </div>
